add support for custom member processing through config directory
